Updated Slide Designs
These slide pngs have the updated logos, designs, and copyrights for 2020. We worked with Kelly to get the resolution increased, so they look okay on large screens. I believe you should be able to replace the image files in the repository with these and continue to use them as you have been.